Camera rotation fixes
When the camera rotates, the main objects rotation stays the same as it was before the rotation and any movements after the camera rotation will affect the rotation correctly. There used to be an issue where the rotation of the main object would not remain the same as it used to be, but would be updated to be the same rotation as it was before in relation to the camera, but now it stays in the same rotation relative to itself.